+++ src/GafferModule/SpreadsheetBinding.cpp
@@ -125,6 +125,8 @@ class RowsPlugSerialiser : public ValuePlugSerialiser
 				return result;
 			}
 
+			// Serialise columns
+
 			for( const auto &cell : Spreadsheet::CellPlug::Range( *plug->getChild<Spreadsheet::RowPlug>( 0 )->cellsPlug() ) )
 			{
 				PlugPtr p = cell->valuePlug()->createCounterpart( cell->getName(), Plug::In );
@@ -137,17 +139,113 @@ class RowsPlugSerialiser : public ValuePlugSerialiser
 				result += " )\n";
 			}
 
+			// Serialise rows. We do this as an `addRows()` call because it is much faster
+			// than serialising a constructor for every single cell. It also shows people the
+			// API they should use for making their own spreadsheets.
+
 			const size_t numRows = plug->children().size();
 			if( numRows > 1 )
 			{
 				result += identifier + ".addRows( " + std::to_string( numRows - 1 ) + " )\n";
 			}
+
+			// If the default values for any cells have been modified, then we need to serialise
+			// those separately.
+
+			std::string defaultValueSerialisation;
+			for( size_t rowIndex = 1; rowIndex < numRows; ++rowIndex )
+			{
+				const auto *row = plug->getChild<Spreadsheet::RowPlug>( rowIndex );
+				defaultValueSerialisationsWalk( row, plug->defaultRow(), serialisation, defaultValueSerialisation );
+			}
+
+			if( defaultValueSerialisation.size() )
+			{
+				result += defaultValueSerialisation + identifier + ".resetDefault()\n";
+			}
+
 			return result;
 		}
 
 		bool childNeedsConstruction( const Gaffer::GraphComponent *child, const Serialisation &serialisation ) const override
 		{
 			// We can serialise much more compactly via the `addRows()` call made by `postConstructor()`.
+			return false;
+		}
+
+	private :
+
+		// Fills `result` with calls to restore default values that have been modified from those of the
+		// default row. We must fit the following constraints :
+		//
+		// - We want to serialise values as high in the hierarchy as we can. For instance,
+		//   a single call at the level of a V3fPlug rather than a call for each of the
+		//   three children. This keeps serialisations smaller and makes loading quicker.
+		// - We want to avoid any redundant work for children if we will serialise at the
+		//   parent level.
+		// - We must have absolutely minimal overhead for the common case where the default
+		//   values have not been modified.
+		//
+		// The strategy is a recursion where each child returns `true` if it needs the caller
+		// the emit a serialisation on its behalf and `false` otherwise.
+		bool defaultValueSerialisationsWalk( const ValuePlug *plug, const ValuePlug *defaultPlug, const Serialisation &serialisation, std::string &result ) const
+		{
+			const size_t numChildren = plug->children().size();
+			assert( defaultPlug->children().size() == numChildren );
+			if( !numChildren )
+			{
+				// Leaf plug. If the default doesn't match the default row then we will
+				// need serialisation. We use `defaultHash()` to avoid the expensive entry
+				// into Python that would be needed to compare values directly.
+				return plug->defaultHash() != defaultPlug->defaultHash();
+			}
+
+			// Compound plug. See if the children need their default values to be serialised
+			// and act accordingly.
+
+			std::vector<const ValuePlug *> childrenToSerialise;
+			for( size_t childIndex = 0; childIndex < numChildren; ++childIndex )
+			{
+				const auto *childPlug = plug->getChild<ValuePlug>( childIndex );
+				const bool childRequiresSerialisation = defaultValueSerialisationsWalk(
+					childPlug, defaultPlug->getChild<ValuePlug>( childIndex ), serialisation, result
+				);
+				if( childRequiresSerialisation )
+				{
+					childrenToSerialise.push_back( childPlug );
+				}
+			}
+
+			if( !childrenToSerialise.size() )
+			{
+				return false;
+			}
+
+			if( childrenToSerialise.size() == numChildren )
+			{
+				// All children want serialisation. As long as we have the appropriate
+				// methods, we can delegate all the work to our parent.
+				object pythonPlug( PlugPtr( const_cast<ValuePlug *>( plug ) ) );
+				if( PyObject_HasAttrString( pythonPlug.ptr(), "setValue" ) )
+				{
+					return true;
+				}
+			}
+
+			// Only a subset of children want to change their default value, or
+			// it's not possible to change the default at this level. Add serialisations
+			// for each child.
+
+			for( auto childPlug : childrenToSerialise )
+			{
+				object pythonChildPlug( PlugPtr( const_cast<ValuePlug *>( childPlug ) ) );
+				object pythonDefaultValue = pythonChildPlug.attr( "defaultValue" )();
+				/// \todo Build identifier recursively (but lazily) and making sure to use the faster
+				/// version of `childIdentifier()`.
+				const std::string childPlugIdentifier = serialisation.identifier( childPlug );
+				result += childPlugIdentifier + ".setValue( " + valueRepr( pythonDefaultValue ) + " )\n";
+			}
+
 			return false;
 		}